Python: Header position should be 64-bits long
Fixes MB-7347

Reviewed-on: http://review.couchbase.org/29590
Tested-by: buildbot <build@couchbase.com>

+++ python/couchstore.py
@@ -103,7 +103,7 @@ class DbInfoStruct (Structure):
                 ("doc_count", c_ulonglong),
                 ("deleted_count", c_ulonglong),
                 ("space_used", c_ulonglong),
-                ("header_position", c_size_t) ]
+                ("header_position", c_ulonglong) ]
 
 
 ### DOCUMENT INFO CLASS:
